more cleanup and renaming
